+++ plugins/healthcheck/CHANGELOG.md
@@ -1,4 +1,7 @@
 # Changelog
 
+## Unreleased
+ - export getHealthCheckResult for easy integration in APM tools
+
 ## 1.0.0
  - Initial release